requirements.txt
# Core
Flask>=2.3.2
Flask-WTF>=1.2.0
Werkzeug>=2.3.4
python-dotenv>=1.0.0

# Geospatial
rasterio>=1.3.5
pyproj>=3.6.0
geopandas>=0.14.0
shapely>=2.0.0
pandas>=2.0.0

# HTTP
requests>=2.31.0

# Routing (Valhalla)
polyline>=2.0.0

# NL-to-GIS (LLM providers — install at least one)
anthropic>=0.40.0
google-genai>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-playwright>=0.7.0
